// src/prisma/migrations/snapshots/2a5f3b3f81fa36854422aa20fcfb20216ec79bf57ff8fbe2b6638a3b08ed60e0/contract.d.ts
export type StorageHash =
  StorageHashBase<'2a5f3b3f81fa36854422aa20fcfb20216ec79bf57ff8fbe2b6638a3b08ed60e0'>;
export type ExecutionHash = ExecutionHashBase<string>;
export type ProfileHash =
  ProfileHashBase<'3916f444a8a17ad749191acf9e08dad97d1a327b88c2f1d45d12f240296aa8b2'>;

export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly DisabledUserBot: { readonly UserId: CodecTypes['pg/text@1']['output'] };
    readonly Group: { readonly GroupId: CodecTypes['pg/text@1']['output'] };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly DisabledUserBot: { readonly UserId: CodecTypes['pg/text@1']['input'] };
    readonly Group: { readonly GroupId: CodecTypes['pg/text@1']['input'] };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly DisabledUserBot: { readonly UserId: CodecTypes['pg/text@1']['output'] };
    readonly Group: { readonly GroupId: CodecTypes['pg/text@1']['output'] };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly DisabledUserBot: { readonly UserId: CodecTypes['pg/text@1']['input'] };
    readonly Group: { readonly GroupId: CodecTypes['pg/text@1']['input'] };
  };
};

export namespace Models {
  export type public_DisabledUserBot = {
    UserId: CodecTypes['pg/text@1']['output'];
    readonly [RelationKeys]?: never;
  };
  export type public_Group = {
    GroupId: CodecTypes['pg/text@1']['output'];
    readonly [RelationKeys]?: never;
  };
}

export declare const models: {
  public: {
    DisabledUserBot: Models.public_DisabledUserBot;
    Group: Models.public_Group;
  };
};

export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;

type ContractBase = Omit<
  ContractType<{
    readonly namespaces: {
      readonly public: {
        readonly id: 'public';
        readonly kind: 'postgres-schema';
        readonly entries: {
          readonly table: {
            readonly DisabledUserBot: {
              columns: {
                readonly UserId: {
                  readonly nativeType: 'text';
                  readonly codecId: 'pg/text@1';
                  readonly nullable: false;
                };
              };
              primaryKey: { readonly columns: readonly ['UserId'] };
              uniques: readonly [];
              indexes: readonly [];
              foreignKeys: readonly [];
            };
            readonly Group: {
              columns: {
                readonly GroupId: {
                  readonly nativeType: 'text';
                  readonly codecId: 'pg/text@1';
                  readonly nullable: false;
                };
              };
              primaryKey: { readonly columns: readonly ['GroupId'] };
              uniques: readonly [];
              indexes: readonly [];
              foreignKeys: readonly [];
            };
          };
        };
      };
    };
    readonly storageHash: StorageHash;
  }>,
  'roots' | 'domain'
> & {
  readonly target: 'postgres';
  readonly targetFamily: 'sql';
  readonly roots: {
    readonly DisabledUserBot: {
      readonly namespace: 'public' & NamespaceId;
      readonly model: 'DisabledUserBot';
    };
    readonly Group: { readonly namespace: 'public' & NamespaceId; readonly model: 'Group' };
  };
  readonly domain: {
    readonly namespaces: {
      readonly public: {
        readonly models: {
          readonly DisabledUserBot: {
            readonly fields: {
              readonly UserId: {
                readonly nullable: false;
                readonly type: { readonly kind: 'scalar'; readonly codecId: 'pg/text@1' };
              };
            };
            readonly relations: Record<string, never>;
            readonly storage: {
              readonly table: 'DisabledUserBot';
              readonly namespaceId: 'public';
              readonly fields: { readonly UserId: { readonly column: 'UserId' } };
            };
          };
          readonly Group: {
            readonly fields: {
              readonly GroupId: {
                readonly nullable: false;
                readonly type: { readonly kind: 'scalar'; readonly codecId: 'pg/text@1' };
              };
            };
            readonly relations: Record<string, never>;
            readonly storage: {
              readonly table: 'Group';
              readonly namespaceId: 'public';
              readonly fields: { readonly GroupId: { readonly column: 'GroupId' } };
            };
          };
        };
      };
    };
  };
  readonly capabilities: {
    readonly postgres: {
      readonly distinctOn: true;
      readonly jsonAgg: true;
      readonly lateral: true;
      readonly limit: true;
      readonly orderBy: true;
      readonly returning: true;
    };
    readonly sql: {
      readonly checkConstraint: true;
      readonly defaultInInsert: true;
      readonly enums: true;
      readonly insertOnConflictSkip: true;
      readonly insertOnConflictWithoutTarget: true;
      readonly lateral: true;
      readonly returning: true;
      readonly scalarList: true;
    };
  };
  readonly extensions: {};
  readonly meta: {};

  readonly profileHash: ProfileHash;
};

export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;